+++ src/app/volunteers/page.tsx
@@ -1,40 +1,5 @@
 // Page that will eventually contain a list of all volunteers
-import VolunteerInfoComponent from '@/components/volunteers/Volunteer';
-import { VolunteerResponse } from '@/types/dataModel/volunteer';
-
-// test case to call VolunteerInfoComponent
-const sampleVolunteer: VolunteerResponse = {
-  firstName: 'John',
-  lastName: 'Doe',
-  email: 'john.doe@example.com',
-  phoneNumber: '123-456-7890',
-  address: '123 Main St, Anytown, USA',
-  softDelete: false,
-  previousOrganization: {
-    name: 'Previous Organization',
-    softDelete: false,
-    _id: 'organization_id',
-    createdAt: new Date(),
-    updatedAt: new Date(),
-  },
-  _id: 'volunteer_id',
-  createdAt: new Date(),
-  updatedAt: new Date(),
-  roleVerifications: [
-    {
-      role: 'Medical',
-      verifier: 'Verifier 1',
-      lastUpdated: new Date(),
-    },
-  ],
-};
 
 export default function VolunteerPage() {
-  // VolunteerInfoComponent with the dummy object
-  return (
-    <h1>
-      {' '}
-      Volunteer Page <VolunteerInfoComponent volunteer={sampleVolunteer} />{' '}
-    </h1>
-  );
+  return <h1> Volunteer Page </h1>;
 }
